decisionai_plugin/requirements.txt
pyyaml
Flask==1.1.1
flask_restful
requests
python-dateutil
azure-storage-blob==12.1.0
azure-cosmosdb-table==1.0.6
Werkzeug==0.16.0
gunicorn==19.9.0
gevent==1.4.0
apscheduler
numpy==1.17.3
pandas==0.23.4

https://pythonpackagerepo.blob.core.windows.net/pip-package/anomaly_detector_telemetry-3.0.tar.gz